requirements.txt
asgiref==3.5.2
astroid==2.11.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
bandit==1.7.4
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
dill==0.3.5.1
Django==4.0.5
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
Flask==2.1.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Markdown==3.3.7
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
pbr==5.9.0
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pylint==2.13.9
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.1.2
wrapt==1.14.1
